+++ tests/HelperNumberTrait/HelperNumberFromStringTest.php
@@ -29,8 +29,13 @@ final class HelperNumberFromStringTest extends TestCase
         $integer = Helper::numberFromString('минус миллион один');
         $this->assertSame($integer, -1000001);
 
+        $integer = Helper::numberFromString('-123');
+        $this->assertSame($integer, -123);
+
+        $integer = Helper::numberFromString(123.456);
+        $this->assertSame($integer, 123.456);
+
         $integer = Helper::numberFromString(null);
-        $this->assertEquals($integer, 0);
         $this->assertTrue($integer === 0);
     }
 }